Fix compiler warnings for openssl 3.0 and curl 7.85
The new compiler gcc-13 is flagging warnings in addition to openssl 3.0
and curl 7.85.
Fix the compiler warnings such as std::min requiring the same parameter
types.
Fix the openssl warnings by replacing the deprecated APIs if version is
3.0 or above.
Fix the curl warnings by replacing the deprecated APIs if version is
7.85 or above.

Tested-by: Matthew Vaught <matthew.vaught@ibm.com>

- RHEL8 with openssl 1.1 and libcurl 7.61.1
- MacOS with openssl 1.1, 3.0 and libcurl 7.87.0

+++ src/client-c++/sf_pkcs11/src/pkcs11-sf-slot.cpp
@@ -50,13 +50,13 @@ CK_FLAGS PKCS11_SfSlot::getFlags() const
 void PKCS11_SfSlot::getDescription(CK_UTF8CHAR* dstParm, uint64_t sizeParm) const
 {
     memset(dstParm, ' ', sizeParm);
-    memcpy(dstParm, mDescription.c_str(), std::min(sizeParm, mDescription.size()));
+    memcpy(dstParm, mDescription.c_str(), std::min(sizeParm, static_cast<uint64_t>(mDescription.size())));
 }
 
 void PKCS11_SfSlot::getManufacturerId(CK_UTF8CHAR* dstParm, uint64_t sizeParm) const
 {
     memset(dstParm, ' ', sizeParm);
-    memcpy(dstParm, mManufacturerId.c_str(), std::min(sizeParm, mManufacturerId.size()));
+    memcpy(dstParm, mManufacturerId.c_str(), std::min(sizeParm, static_cast<uint64_t>(mManufacturerId.size())));
 }
 
 CK_VERSION PKCS11_SfSlot::getHardwareVersion() const { return mHardwareVersion; }
